+++ docker/atari.Dockerfile
@@ -1,5 +1,11 @@
 FROM nvcr.io/nvidia/jax:23.10-py3
 
+# run stuff as non-root, comment if you want to run as root
+ARG UID
+ARG USERNAME
+RUN useradd -u $UID --create-home $USERNAME
+USER $USERNAME
+
 # install 
 WORKDIR /app/
 COPY . .
