+++ docker/run.sh
@@ -1,7 +1,9 @@
-gpu=0 # you can set this to 0,1,2 for multiple gpus or "all" for all gpus
-docker run -it --rm --gpus=\"device=${gpu}\" --ipc=host \
+gpu=1 # you can set this to 0,1,2 for multiple gpus or "all" for all gpus
+docker run -d --rm --gpus=\"device=${gpu}\" --ipc=host \
  -v $(pwd):/app \
  -w /app \
  --name "pqn-${gpu//,/-}" \
  pqn \
- bash
+ bash # python purejaxql/pqn_vdn_rnn_jaxmarl.py +alg=pqn_vdn_rnn_smax
+ #python purejaxql/pqn_craftax.py +alg=pqn_craftax
+ #python purejaxql/pqn_vdn_rnn_jaxmarl.py +alg=pqn_vdn_rnn_smax
