--- purejaxql/pqn_brax.py
@@ -1,469 +1,0 @@
-import os
-import time
-import copy
-import jax
-import jax.numpy as jnp
-import flax.linen as nn
-import numpy as np
-import optax
-from flax.linen.initializers import constant, orthogonal
-from typing import Sequence, NamedTuple, Any
-from flax.training.train_state import TrainState
-import distrax
-import wandb
-import hydra
-from omegaconf import OmegaConf
-from wrappers import (
-    LogWrapper,
-    BraxGymnaxWrapper,
-    VecEnv,
-    NormalizeVecObservation,
-    NormalizeVecReward,
-    ClipAction,
-)
-
-
-class Actor(nn.Module):
-    action_dim: Sequence[int]
-    activation: str = "tanh"
-
-    @nn.compact
-    def __call__(self, x):
-        if self.activation == "relu":
-            activation = nn.relu
-        else:
-            activation = nn.tanh
-        x = nn.Dense(
-            256, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0)
-        )(x)
-        x = nn.LayerNorm()(x)
-        x = activation(x)
-        x = nn.Dense(
-            256, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0)
-        )(x)
-        x = nn.LayerNorm()(x)
-        x = activation(x)
-        x = nn.Dense(
-            self.action_dim, kernel_init=orthogonal(0.01), bias_init=constant(0.0)
-        )(x)
-        actor_logtstd = self.param("log_std", nn.initializers.zeros, (self.action_dim,))
-        pi = distrax.MultivariateNormalDiag(x, jnp.exp(actor_logtstd))
-        return pi
-    
-
-class Critic(nn.Module):
-    @nn.compact
-    def __call__(self, x, action):
-        x = jnp.concatenate([x, action], axis=-1)
-        x = nn.Dense(
-            256, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0)
-        )(x)
-        x = nn.LayerNorm()(x)
-        x = nn.relu(x)
-        x = nn.Dense(
-            256, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0)
-        )(x)
-        x = nn.LayerNorm()(x)
-        x = nn.relu(x)
-        x = nn.Dense(1, kernel_init=orthogonal(1.0), bias_init=constant(0.0))(
-            x
-        )
-        return jnp.squeeze(x, axis=-1)
-
-
-class ActorCritic(nn.Module):
-    action_dim: Sequence[int]
-    activation: str = "tanh"
-
-    @nn.compact
-    def __call__(self, x, rng):
-        
-        pi = Actor(self.action_dim, self.activation, name='actor')(x)
-        action = pi.sample(seed=rng)
-        value = Critic(name='critic')(x, action)
-
-        return pi, value, action
-        
-
-
-class Transition(NamedTuple):
-    done: jnp.ndarray
-    action: jnp.ndarray
-    value: jnp.ndarray
-    reward: jnp.ndarray
-    log_prob: jnp.ndarray
-    obs: jnp.ndarray
-    info: jnp.ndarray
-
-
-class CustomTrainState(TrainState):
-    timesteps: int = 0
-    n_updates: int = 0
-    grad_steps: int = 0
-
-
-def make_train(config):
-    config["NUM_UPDATES"] = (
-        config["TOTAL_TIMESTEPS"] // config["NUM_STEPS"] // config["NUM_ENVS"]
-    )
-    config["MINIBATCH_SIZE"] = (
-        config["NUM_ENVS"] * config["NUM_STEPS"] // config["NUM_MINIBATCHES"]
-    )
-    env, env_params = BraxGymnaxWrapper(config["ENV_NAME"]), None
-    env = LogWrapper(env)
-    env = ClipAction(env)
-    env = VecEnv(env)
-    if config["NORMALIZE_ENV"]:
-        env = NormalizeVecObservation(env)
-        env = NormalizeVecReward(env, config["GAMMA"])
-
-    def linear_schedule(count):
-        frac = (
-            1.0
-            - (count // (config["NUM_MINIBATCHES"] * config["UPDATE_EPOCHS"]))
-            / config["NUM_UPDATES"]
-        )
-        return config["LR"] * frac
-
-    def train(rng):
-
-        original_rng = rng[0]
-
-        # INIT NETWORK
-        actor = Actor(env.action_space(env_params).shape[0], activation=config["ACTIVATION"])
-        critic = Critic()
-        network = ActorCritic(
-            env.action_space(env_params).shape[0], activation=config["ACTIVATION"]
-        )
-        rng, _rng = jax.random.split(rng)
-        init_x = jnp.zeros(env.observation_space(env_params).shape)
-        network_params = network.init(_rng, init_x, rng)
-        if config["ANNEAL_LR"]:
-            tx = optax.chain(
-                optax.clip_by_global_norm(config["MAX_GRAD_NORM"]),
-                optax.adam(learning_rate=linear_schedule, eps=1e-5),
-            )
-        else:
-            tx = optax.chain(
-                optax.clip_by_global_norm(config["MAX_GRAD_NORM"]),
-                optax.adam(config["LR"], eps=1e-5),
-            )
-        train_state = CustomTrainState.create(
-            apply_fn=network.apply,
-            params=network_params,
-            tx=tx,
-        )
-
-        # INIT ENV
-        rng, _rng = jax.random.split(rng)
-        reset_rng = jax.random.split(_rng, config["NUM_ENVS"])
-        obsv, env_state = env.reset(reset_rng, env_params)
-
-        # TRAIN LOOP
-        def _update_step(runner_state, unused):
-            # COLLECT TRAJECTORIES
-            def _env_step(runner_state, unused):
-                train_state, env_state, last_obs, rng = runner_state
-
-                # SELECT ACTION
-                rng, _rng = jax.random.split(rng)
-                pi, value, action = network.apply(train_state.params, last_obs, _rng)
-                log_prob = pi.log_prob(action)
-
-                # STEP ENV
-                rng, _rng = jax.random.split(rng)
-                rng_step = jax.random.split(_rng, config["NUM_ENVS"])
-                obsv, env_state, reward, done, info = env.step(
-                    rng_step, env_state, action, env_params
-                )
-                transition = Transition(
-                    done, action, value, reward, log_prob, last_obs, info
-                )
-                runner_state = (train_state, env_state, obsv, rng)
-                return runner_state, transition
-
-            runner_state, traj_batch = jax.lax.scan(
-                _env_step, runner_state, None, config["NUM_STEPS"]
-            )
-
-            # CALCULATE ADVANTAGE
-            train_state, env_state, last_obs, rng = runner_state
-            _, last_val, _ = network.apply(train_state.params, last_obs, _rng)
-
-            def _get_target(lambda_returns_and_next_q, transition):
-                lambda_returns, next_q = lambda_returns_and_next_q
-                target_bootstrap = (
-                    transition.reward + config["GAMMA"] * (1 - transition.done) * next_q
-                )
-                delta = lambda_returns - next_q
-                lambda_returns = (
-                    target_bootstrap + config["GAMMA"] * config["LAMBDA"] * delta
-                )
-                lambda_returns = (
-                    1 - transition.done
-                ) * lambda_returns + transition.done * transition.reward
-                next_q = transition.value
-                return (lambda_returns, next_q), lambda_returns
-
-            last_val = last_val * (1 - traj_batch.done[-1])
-            lambda_returns = traj_batch.reward[-1] + config["GAMMA"] * last_val
-            _, targets = jax.lax.scan(
-                _get_target,
-                (lambda_returns, last_val),
-                jax.tree_map(lambda x: x[:-1], traj_batch),
-                reverse=True,
-            )
-            targets = jnp.concatenate((targets, lambda_returns[np.newaxis]))
-            advantages = (targets - targets.mean()) / (targets.std() + 1e-8)
-
-            # UPDATE NETWORK
-            def _update_epoch(update_state, unused):
-                def _update_minbatch(train_state, batch_info):
-                    traj_batch, advantages, targets = batch_info
-
-                    def _loss_fn(params, traj_batch, gae, targets):
-
-                        # CALCULATE VALUE LOSS
-                        value = critic.apply({'params':params['params']['critic']}, traj_batch.obs, traj_batch.action)
-                        """
-                        value_pred_clipped = traj_batch.value + (
-                            value - traj_batch.value
-                        ).clip(-config["CLIP_EPS"], config["CLIP_EPS"])
-                        value_losses = jnp.square(value - targets)
-                        value_losses_clipped = jnp.square(value_pred_clipped - targets)
-                        value_loss = (
-                            0.5 * jnp.maximum(value_losses, value_losses_clipped).mean()
-                        )
-                        """
-                        value_loss = 0.5 * jnp.square(value - targets).mean()
-
-                        # CALCULATE ACTOR LOSS
-                        pi = actor.apply({'params':params['params']['actor']}, traj_batch.obs)
-                        log_prob = pi.log_prob(traj_batch.action)
-
-                        # ppo clip loss
-                        ratio = jnp.exp(log_prob - traj_batch.log_prob)
-                        #gae = (gae - gae.mean()) / (gae.std() + 1e-8)
-                        loss_actor1 = ratio * gae
-                        loss_actor2 = (
-                            jnp.clip(
-                                ratio,
-                                1.0 - config["CLIP_EPS"],
-                                1.0 + config["CLIP_EPS"],
-                            )
-                            * gae
-                        )
-                        loss_actor = -jnp.minimum(loss_actor1, loss_actor2)
-                        #ratio = jnp.exp(log_prob - traj_batch.log_prob)
-                        #loss_actor = -(ratio * gae)
-                        loss_actor = loss_actor.mean()
-                        entropy = pi.entropy().mean()
-
-                        total_loss = (
-                            loss_actor
-                            + config["VF_COEF"] * value_loss
-                            - config["ENT_COEF"] * entropy
-                        )
-                        return total_loss, (value_loss, loss_actor, entropy)
-
-                    grad_fn = jax.value_and_grad(_loss_fn, has_aux=True)
-                    (total_loss, (value_loss, loss_actor, entropy)), grads = grad_fn(
-                        train_state.params, traj_batch, advantages, targets
-                    )
-                    train_state = train_state.apply_gradients(grads=grads)
-                    train_state = train_state.replace(grad_steps=train_state.grad_steps + 1)
-                    return train_state, (total_loss, (value_loss, loss_actor, entropy))
-
-                train_state, traj_batch, advantages, targets, rng = update_state
-                rng, _rng = jax.random.split(rng)
-                batch_size = config["MINIBATCH_SIZE"] * config["NUM_MINIBATCHES"]
-                assert (
-                    batch_size == config["NUM_STEPS"] * config["NUM_ENVS"]
-                ), "batch size must be equal to number of steps * number of envs"
-                permutation = jax.random.permutation(_rng, batch_size)
-                batch = (traj_batch, advantages, targets)
-                batch = jax.tree_util.tree_map(
-                    lambda x: x.reshape((batch_size,) + x.shape[2:]), batch
-                )
-                shuffled_batch = jax.tree_util.tree_map(
-                    lambda x: jnp.take(x, permutation, axis=0), batch
-                )
-                minibatches = jax.tree_util.tree_map(
-                    lambda x: jnp.reshape(
-                        x, [config["NUM_MINIBATCHES"], -1] + list(x.shape[1:])
-                    ),
-                    shuffled_batch,
-                )
-                train_state, loss_info = jax.lax.scan(
-                    _update_minbatch, train_state, minibatches
-                )
-                update_state = (train_state, traj_batch, advantages, targets, rng)
-                return update_state, loss_info
-
-            update_state = (train_state, traj_batch, advantages, targets, rng)
-            update_state, (total_loss, (value_loss, loss_actor, entropy)) = jax.lax.scan(
-                _update_epoch, update_state, None, config["UPDATE_EPOCHS"]
-            )
-            train_state = update_state[0]
-            metric = traj_batch.info
-            rng = update_state[-1]
-
-            train_state = train_state.replace(
-                timesteps=(train_state.n_updates + 1)*config["NUM_ENVS"]*config["NUM_STEPS"],
-                n_updates=train_state.n_updates + 1,
-            )
-            metrics = {
-                "env_step": train_state.timesteps,
-                "update_steps": train_state.n_updates,
-                "grad_steps": train_state.grad_steps,
-                "loss": total_loss.mean(),
-                "value_loss": value_loss.mean(),
-                "loss_actor": loss_actor.mean(),
-                "entropy": entropy.mean(),
-            }
-            metrics.update({k: v.mean() for k, v in traj_batch.info.items()})
-
-            # report on wandb if required
-            if config["WANDB_MODE"] != "disabled":
-
-                def callback(metrics, original_rng):
-                    if config.get("WANDB_LOG_ALL_SEEDS", False):
-                        metrics.update(
-                            {
-                                f"rng{int(original_rng)}/{k}": v
-                                for k, v in metrics.items()
-                            }
-                        )
-                    wandb.log(metrics, step=metrics["update_steps"])
-
-                jax.debug.callback(callback, metrics, original_rng)
-
-            runner_state = (train_state, env_state, last_obs, rng)
-            return runner_state, metric
-
-        rng, _rng = jax.random.split(rng)
-        runner_state = (train_state, env_state, obsv, _rng)
-        runner_state, metric = jax.lax.scan(
-            _update_step, runner_state, None, config["NUM_UPDATES"]
-        )
-        return {"runner_state": runner_state, "metrics": metric}
-
-    return train
-
-
-def single_run(config):
-
-    config = {**config, **config["alg"]}
-
-    alg_name = config.get("ALG_NAME", "ppo")
-    env_name = config["ENV_NAME"]
-
-    wandb.init(
-        entity=config["ENTITY"],
-        project=config["PROJECT"],
-        tags=[
-            alg_name.upper(),
-            env_name.upper(),
-            f"jax_{jax.__version__}",
-        ],
-        name=f'{config["ALG_NAME"]}_{config["ENV_NAME"]}',
-        config=config,
-        mode=config["WANDB_MODE"],
-    )
-
-    rng = jax.random.PRNGKey(config["SEED"])
-
-    t0 = time.time()
-    rngs = jax.random.split(rng, config["NUM_SEEDS"])
-    train_vjit = jax.jit(jax.vmap(make_train(config)))
-    outs = jax.block_until_ready(train_vjit(rngs))
-    print(f"Took {time.time()-t0} seconds to complete.")
-
-    if config.get("SAVE_PATH", None) is not None:
-        from jaxmarl.wrappers.baselines import save_params
-
-        model_state = outs["runner_state"][0]
-        save_dir = os.path.join(config["SAVE_PATH"], env_name)
-        os.makedirs(save_dir, exist_ok=True)
-        OmegaConf.save(
-            config,
-            os.path.join(
-                save_dir, f'{alg_name}_{env_name}_seed{config["SEED"]}_config.yaml'
-            ),
-        )
-
-        for i, rng in enumerate(rngs):
-            params = jax.tree_map(lambda x: x[i], model_state.params)
-            save_path = os.path.join(
-                save_dir,
-                f'{alg_name}_{env_name}_seed{config["SEED"]}_vmap{i}.safetensors',
-            )
-            save_params(params, save_path)
-
-
-def tune(default_config):
-    """Hyperparameter sweep with wandb."""
-
-    default_config = {**default_config, **default_config["alg"]}
-    alg_name = default_config.get("ALG_NAME", "ppo")
-    env_name = default_config["ENV_NAME"]
-
-    def wrapped_make_train():
-        wandb.init(project=default_config["PROJECT"])
-
-        config = copy.deepcopy(default_config)
-        for k, v in dict(wandb.config).items():
-            config[k] = v
-
-            print("running experiment with params:", config)
-
-            rng = jax.random.PRNGKey(config["SEED"])
-            rngs = jax.random.split(rng, config["NUM_SEEDS"])
-            train_vjit = jax.jit(jax.vmap(make_train(config)))
-            outs = jax.block_until_ready(train_vjit(rngs))
-
-    sweep_config = {
-        "name": f"{alg_name}_{env_name}",
-        "method": "bayes",
-        "metric": {
-            "name": "returned_episode_returns",
-            "goal": "maximize",
-        },
-        "parameters": {
-            "LR": {
-                "values": [
-                    0.0005,
-                    0.0001,
-                    0.00005,
-                    0.00001,
-                    0.000005,
-                ]
-            },
-            "LAMBDA": {"values": [0.95, 0.9, 0.8, 0.7, 0.6]},
-            #"ACTIVATION": {"values": ["tanh", "relu"]},
-            #"ANNEAL_LR": {"values": [True, False]},
-            #"NORMALIZE_ENV": {"values": [True, False]},
-            "NUM_MINIBATCHES": {"values": [4, 8, 16, 32, 64]},
-            #"CLIP_EPS": {"values": [0.1, 0.2]},
-        },
-    }
-
-    wandb.login()
-    sweep_id = wandb.sweep(
-        sweep_config, entity=default_config["ENTITY"], project=default_config["PROJECT"]
-    )
-    wandb.agent(sweep_id, wrapped_make_train, count=1000)
-
-
-@hydra.main(version_base=None, config_path="./config", config_name="config")
-def main(config):
-    config = OmegaConf.to_container(config)
-    print("Config:\n", OmegaConf.to_yaml(config))
-    if config["HYP_TUNE"]:
-        tune(config)
-    else:
-        single_run(config)
-
-
-if __name__ == "__main__":
-    main()
