+++ purejaxql/pqn_rnn_craftax.py
@@ -663,7 +663,7 @@ def single_run(config):
     print(f"Took {time.time()-t0} seconds to complete.")
 
     if config.get("SAVE_PATH", None) is not None:
-        from jaxmarl.wrappers.baselines import save_params
+        from purejaxql.utils.save_load import save_params
 
         model_state = outs["runner_state"][0]
         save_dir = os.path.join(config["SAVE_PATH"], env_name)
