+++ purejaxql/pqn_rnn_gymnax.py
@@ -600,10 +600,7 @@ def single_run(config):
     print(f"Took {time.time()-t0} seconds to complete.")
 
     if config.get("SAVE_PATH", None) is not None:
-        try:
-            from jaxmarl.wrappers.baselines import save_params
-        except ImportError:
-            from purejaxql.pqn_gymnax_flat import save_params
+        from purejaxql.utils.save_load import save_params
 
         model_state = outs["runner_state"][0]
         save_dir = os.path.join(config["SAVE_PATH"], env_name)
