+++ purejaxql/utils/atari_wrapper.py
@@ -1,0 +1,169 @@
+import jax
+import jax.numpy as jnp
+from flax import struct
+import numpy as np
+from functools import partial
+
+import gym
+from packaging import version
+
+is_legacy_gym = version.parse(gym.__version__) < version.parse("0.26.0")
+assert is_legacy_gym, "Current version supports only gym<=0.23.1"
+
+# (random,human)
+ATARI_SCORES = {
+    "Alien-v5": (227.8, 7127.7),
+    "Amidar-v5": (5.8, 1719.5),
+    "Assault-v5": (222.4, 742.0),
+    "Asterix-v5": (210.0, 8503.3),
+    "Asteroids-v5": (719.1, 47388.7),
+    "Atlantis-v5": (12850.0, 29028.1),
+    "BankHeist-v5": (14.2, 753.1),
+    "BattleZone-v5": (2360.0, 37187.5),
+    "BeamRider-v5": (363.9, 16926.5),
+    "Berzerk-v5": (123.7, 2630.4),
+    "Bowling-v5": (23.1, 160.7),
+    "Boxing-v5": (0.1, 12.1),
+    "Breakout-v5": (1.7, 30.5),
+    "Centipede-v5": (2090.9, 12017.0),
+    "ChopperCommand-v5": (811.0, 7387.8),
+    "CrazyClimber-v5": (10780.5, 35829.4),
+    "Defender-v5": (2874.5, 18688.9),
+    "DemonAttack-v5": (152.1, 1971.0),
+    "DoubleDunk-v5": (-18.6, -16.4),
+    "Enduro-v5": (0.0, 860.5),
+    "FishingDerby-v5": (-91.7, -38.7),
+    "Freeway-v5": (0.0, 29.6),
+    "Frostbite-v5": (65.2, 4334.7),
+    "Gopher-v5": (257.6, 2412.5),
+    "Gravitar-v5": (173.0, 3351.4),
+    "Hero-v5": (1027.0, 30826.4),
+    "IceHockey-v5": (-11.2, 0.9),
+    "Jamesbond-v5": (29.0, 302.8),
+    "Kangaroo-v5": (52.0, 3035.0),
+    "Krull-v5": (1598.0, 2665.5),
+    "KungFuMaster-v5": (258.5, 22736.3),
+    "MontezumaRevenge-v5": (0.0, 4753.3),
+    "MsPacman-v5": (307.3, 6951.6),
+    "NameThisGame-v5": (2292.3, 8049.0),
+    "Phoenix-v5": (761.4, 7242.6),
+    "Pitfall-v5": (-229.4, 6463.7),
+    "Pong-v5": (-20.7, 14.6),
+    "PrivateEye-v5": (24.9, 69571.3),
+    "Qbert-v5": (163.9, 13455.0),
+    "Riverraid-v5": (1338.5, 17118.0),
+    "RoadRunner-v5": (11.5, 7845.0),
+    "Robotank-v5": (2.2, 11.9),
+    "Seaquest-v5": (68.4, 42054.7),
+    "Skiing-v5": (-17098.1, -4336.9),
+    "Solaris-v5": (1236.3, 12326.7),
+    "SpaceInvaders-v5": (148.0, 1668.7),
+    "StarGunner-v5": (664.0, 10250.0),
+    "Surround-v5": (-10.0, 6.5),
+    "Tennis-v5": (-23.8, -8.3),
+    "TimePilot-v5": (3568.0, 5229.2),
+    "Tutankham-v5": (11.4, 167.6),
+    "UpNDown-v5": (533.4, 11693.2),
+    "Venture-v5": (0.0, 1187.5),
+    "VideoPinball-v5": (16256.9, 17667.9),
+    "WizardOfWor-v5": (563.5, 4756.5),
+    "YarsRevenge-v5": (3092.9, 54576.9),
+    "Zaxxon-v5": (32.5, 9173.3),
+}
+
+
+@struct.dataclass
+class LogEnvState:
+    handle: jnp.array
+    lives: jnp.array
+    episode_returns: jnp.array
+    episode_lengths: jnp.array
+    returned_episode_returns: jnp.array
+    returned_episode_lengths: jnp.array
+
+
+class JaxLogEnvPoolWrapper(gym.Wrapper):
+
+    def __init__(self, env, reset_info=True):
+        super(JaxLogEnvPoolWrapper, self).__init__(env)
+        self.num_envs = getattr(env, "num_envs", 1)
+        self.env_name = env.name
+        self.env_random_score, self.env_human_score = ATARI_SCORES[self.env_name]
+        # get if the env has lives
+        self.has_lives = False
+        env.reset()
+        info = env.step(np.zeros(self.num_envs, dtype=int))[-1]
+        if info["lives"].sum() > 0:
+            self.has_lives = True
+            print("env has lives")
+        self.reset_info = reset_info
+        handle, recv, send, step = env.xla()
+        self.init_handle = handle
+        self.send_f = send
+        self.recv_f = recv
+        self.step_f = step
+
+    def reset(self, **kwargs):
+        observations = super(JaxLogEnvPoolWrapper, self).reset(**kwargs)
+
+        env_state = LogEnvState(
+            jnp.array(self.init_handle),
+            jnp.zeros(self.num_envs, dtype=jnp.float32),
+            jnp.zeros(self.num_envs, dtype=jnp.float32),
+            jnp.zeros(self.num_envs, dtype=jnp.float32),
+            jnp.zeros(self.num_envs, dtype=jnp.float32),
+            jnp.zeros(self.num_envs, dtype=jnp.float32),
+        )
+        return observations, env_state
+
+    @partial(jax.jit, static_argnums=(0,))
+    def step(self, state, action):
+
+        new_handle, (observations, rewards, dones, infos) = self.step_f(
+            state.handle, action
+        )
+
+        new_episode_return = state.episode_returns + infos["reward"]
+        new_episode_length = state.episode_lengths + 1
+        state = state.replace(
+            handle=new_handle,
+            episode_returns=(new_episode_return)
+            * (1 - infos["terminated"])
+            * (1 - infos["TimeLimit.truncated"]),
+            episode_lengths=(new_episode_length)
+            * (1 - infos["terminated"])
+            * (1 - infos["TimeLimit.truncated"]),
+            returned_episode_returns=jnp.where(
+                infos["terminated"] + infos["TimeLimit.truncated"],
+                new_episode_return,
+                state.returned_episode_returns,
+            ),
+            returned_episode_lengths=jnp.where(
+                infos["terminated"] + infos["TimeLimit.truncated"],
+                new_episode_length,
+                state.returned_episode_lengths,
+            ),
+        )
+
+        if self.reset_info:
+            elapsed_steps = infos["elapsed_step"]
+            terminated = infos["terminated"] + infos["TimeLimit.truncated"]
+            infos = {}
+        normalize_score = lambda x: (x - self.env_random_score) / (
+            self.env_human_score - self.env_random_score
+        )
+        infos["returned_episode_returns"] = state.returned_episode_returns
+        infos["normalized_returned_episode_returns"] = normalize_score(
+            state.returned_episode_returns
+        )
+        infos["returned_episode_lengths"] = state.returned_episode_lengths
+        infos["elapsed_step"] = elapsed_steps
+        infos["returned_episode"] = terminated
+
+        return (
+            observations,
+            state,
+            rewards,
+            dones,
+            infos,
+        )
