+++ purejaxql/utils/save_load.py
@@ -1,0 +1,12 @@
+import os
+from safetensors.flax import save_file, load_file
+from flax.traverse_util import flatten_dict, unflatten_dict
+from typing import Dict, Union
+
+def save_params(params: Dict, filename: Union[str, os.PathLike]) -> None:
+    flattened_dict = flatten_dict(params, sep=',')
+    save_file(flattened_dict, filename)
+
+def load_params(filename:Union[str, os.PathLike]) -> Dict:
+    flattened_dict = load_file(filename)
+    return unflatten_dict(flattened_dict, sep=",")
